requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# PyYT-MP3/src/funcs.py: 17
# PyYT-MP3/src/main.py: 39
pytube3 == 9.6.4

# PyYT-MP3/src/funcs.py: 11
requests == 2.22.0

# PyYT-MP3/src/funcs.py: 12
# PyYT-MP3/src/main.py: 37
validators == 0.16.0
